pyproject.toml
[tool.poetry]
name = "drf-car-project"
version = "0.1.0"
description = ""
authors = ["egor <ega.nik5674@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.27.2"
django-filter = "^24.3"
python-dotenv = "^1.0.1"
flake8 = "^7.1.1"
black = "^24.8.0"
isort = "^5.13.2"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
gunicorn = "^22.0.0"
coverage = "^7.6.1"
coverage-badge = "^1.1.2"

[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    )/

)
'''

[tool.isort]
line_length = 119

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
